+++ bin/pos-cli-modules-pull.js
@@ -22,7 +22,7 @@ program
     gateway
       .appExportStart({ module_name: module })
       .then(exportTask => {
-        waitForStatus(() => gateway.appExportStatus(exportTask.id))
+        waitForStatus(() => gateway.appExportStatus(exportTask.id), 'ready_for_export', 'success')
           .then(exportTask => downloadFile(exportTask.zip_file.url, filename))
           .then(() => spinner.succeed('Downloading files'))
           .catch(error => {